+++ woninet/core/subnet_enumerator.py
@@ -1,8 +1,8 @@
 from typing import Dict
 from woninet.core.models import Device
-from woninet.utilities.logger import logger_function
+from woninet.utilities.logger import get_core_logger
 
-rootLogger = logger_function()
+core_logger = get_core_logger()
 
 
 class SubnetEnumerator:
@@ -13,7 +13,7 @@ class SubnetEnumerator:
     def scan_subnet(self, ip_addr: str) -> Dict[str, Device]:
         ip_split = ip_addr.split(".")
         subnet = f"{ip_split[0]}.{ip_split[1]}.{ip_split[2]}"
-        rootLogger.debug(f"Building a dictionaty of IP addresses base on {subnet}.x")
+        core_logger.debug(f"Building a dictionaty of IP addresses base on {subnet}.x")
         devices = {}
         for i in range(1, 255):
             ip = f"{subnet}.{i}"